refactoring: 통일을 위해 패키지명(폴더명) 변경 및 이동
