+++ backend/src/main/java/com/dwbh/backend/dto/counsel_offer/CreateOrUpdateOfferRequest.java
@@ -1,4 +1,4 @@
-package com.dwbh.backend.dto.offer;
+package com.dwbh.backend.dto.counsel_offer;
 
 import com.dwbh.backend.common.entity.YnType;
 import io.swagger.v3.oas.annotations.media.Schema;
