+++ backend/src/main/java/com/dwbh/backend/repository/counsel_offer/CounselOfferRepository.java
@@ -2,6 +2,15 @@ package com.dwbh.backend.repository.counsel_offer;
 
 import com.dwbh.backend.entity.CounselOffer;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface CounselOfferRepository extends JpaRepository<CounselOffer, Long> {
+import java.time.LocalDateTime;
+
+public interface CounselOfferRepository extends JpaRepository<CounselOffer, Long>, OfferCustomRepository  {
+
+    @Modifying
+    @Query("UPDATE CounselOffer o SET o.delDate = :delDate WHERE o.offerSeq = :offerSeq")
+    void softDeleteById(@Param("offerSeq") Long offerSeq, @Param("delDate") LocalDateTime delDate);
 }
